player util
