+++ DeFuge/AppDelegate.swift
@@ -12,33 +12,15 @@ import AudioToolbox
 import AVFoundation
 import UserNotifications
 
-protocol AlarmApplicationDelegate {
-    func playMusic(_ soundName: String)
-}
+
 @UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate,AVAudioPlayerDelegate,UNUserNotificationCenterDelegate {
+class AppDelegate: UIResponder, UIApplicationDelegate,UNUserNotificationCenterDelegate {
 
     var window: UIWindow?
     var audioPlayer: AVAudioPlayer?
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
-        
-        //AVplayer
-        var error: NSError?
-        do {
-            try AVAudioSession.sharedInstance().setCategory(AVAudioSessionCategoryPlayback)
-        } catch let error1 as NSError{
-            error = error1
-            print("could not set session. err:\(error!.localizedDescription)")
-        }
-        do {
-            try AVAudioSession.sharedInstance().setActive(true)
-        } catch let error1 as NSError{
-            error = error1
-            print("could not active session. err:\(error!.localizedDescription)")
-        }
-        window?.tintColor = UIColor.red
         
         // TODO: determine if alarm view is to be shown
         // set true to show alarm view
@@ -55,49 +37,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate,AVAudioPlayerDelegate,UNUs
         }
         
         return true
-    }
-
-    func playMusic(_ soundName: String) {
-        
-        //vibrate phone first
-        AudioServicesPlaySystemSound(SystemSoundID(kSystemSoundID_Vibrate))
-        //set vibrate callback
-        AudioServicesAddSystemSoundCompletion(SystemSoundID(kSystemSoundID_Vibrate),nil,
-                                              nil,
-                                              { (_:SystemSoundID, _:UnsafeMutableRawPointer?) -> Void in
-                                                AudioServicesPlaySystemSound(SystemSoundID(kSystemSoundID_Vibrate))
-        },
-                                              nil)
-        let url = URL(fileURLWithPath: Bundle.main.path(forResource: soundName, ofType: "mp3")!)
-        
-        var error: NSError?
-        
-        do {
-            audioPlayer = try AVAudioPlayer(contentsOf: url)
-        } catch let error1 as NSError {
-            error = error1
-            audioPlayer = nil
-        }
-        
-        if let err = error {
-            print("audioPlayer error \(err.localizedDescription)")
-            return
-        } else {
-            audioPlayer!.delegate = self
-            audioPlayer!.prepareToPlay()
-        }
-        
-        //negative number means loop infinity
-        audioPlayer!.numberOfLoops = -1
-        audioPlayer!.play()
-    }
-    
-        func audioPlayerDidFinishPlaying(_ player: AVAudioPlayer, successfully flag: Bool) {
-        
-    }
-    
-    func audioPlayerDecodeErrorDidOccur(_ player: AVAudioPlayer, error: Error?) {
-        
     }
 
     func applicationWillResignActive(_ application: UIApplication) {
